chore: Moved validator to utils
